Make the schema iterators an internal concept of `SchemaFrame`
