+++ src/frame/include/sourcemeta/blaze/frame_iterator.h
@@ -1,60 +1,67 @@
-#ifndef SOURCEMETA_BLAZE_FOUNDATION_WALKER_H_
-#define SOURCEMETA_BLAZE_FOUNDATION_WALKER_H_
+#ifndef SOURCEMETA_BLAZE_FRAME_ITERATOR_H_
+#define SOURCEMETA_BLAZE_FRAME_ITERATOR_H_
 
-#ifndef SOURCEMETA_BLAZE_FOUNDATION_EXPORT
-#include <sourcemeta/blaze/foundation_export.h>
+#ifndef SOURCEMETA_BLAZE_FRAME_EXPORT
+#include <sourcemeta/blaze/frame_export.h>
 #endif
 
 #include <sourcemeta/core/json.h>
 
-#include <sourcemeta/blaze/foundation_types.h>
+#include <sourcemeta/blaze/foundation.h>
 
 #include <string_view> // std::string_view
 #include <vector>      // std::vector
 
 namespace sourcemeta::blaze {
 
-/// @ingroup foundation
+/// @ingroup frame
 ///
-/// Return an iterator over the top-level keywords of a given JSON Schema
-/// definition in the order in which an implementation must evaluate them,
-/// given the vocabularies in use by the schema.
+/// Return an iterator over the subschemas of a given JSON Schema definition
+/// according to the applicators understood by the provided walker function.
+/// This walker recursively traverses over every subschema of
+/// the JSON Schema definition, including the top-level schema, reporting back
+/// each subschema.
 ///
 /// For example:
 ///
 /// ```cpp
 /// #include <sourcemeta/core/json.h>
 /// #include <sourcemeta/blaze/foundation.h>
+/// #include <sourcemeta/blaze/frame.h>
 /// #include <iostream>
 ///
 /// const sourcemeta::core::JSON document =
 ///   sourcemeta::core::parse_json(R"JSON({
 ///   "$schema": "https://json-schema.org/draft/2020-12/schema",
 ///   "type": "object",
-///   "properties": {},
-///   "additionalProperties": true,
-///   "patternProperties": {}
+///   "properties": {
+///     "foo": {
+///       "type": "array",
+///       "items": {
+///         "type": "string"
+///       }
+///     }
+///   }
 /// })JSON");
 ///
-/// const auto vocabularies{sourcemeta::blaze::vocabularies(
-///     document, sourcemeta::blaze::schema_resolver)};
-///
 /// for (const auto &entry :
-///          sourcemeta::blaze::SchemaKeywordIterator{
-///          document, sourcemeta::blaze::schema_walker, vocabularies}) {
-///   sourcemeta::core::stringify(entry.pointer, std::cout);
+///          sourcemeta::blaze::SchemaIterator{
+///          document, sourcemeta::blaze::schema_walker,
+///          sourcemeta::blaze::schema_resolver}) {
+///   sourcemeta::core::prettify(
+///     sourcemeta::core::get(document, entry.pointer), std::cout);
 ///   std::cout << "\n";
 /// }
 /// ```
-class SOURCEMETA_BLAZE_FOUNDATION_EXPORT SchemaKeywordIterator {
+class SOURCEMETA_BLAZE_FRAME_EXPORT SchemaIterator {
 private:
   using internal = typename std::vector<SchemaIteratorEntry>;
 
 public:
   using const_iterator = typename internal::const_iterator;
-  SchemaKeywordIterator(const sourcemeta::core::JSON &input,
-                        const SchemaWalker &walker,
-                        const Vocabularies &vocabularies);
+  SchemaIterator(const sourcemeta::core::JSON &input,
+                 const SchemaWalker &walker, const SchemaResolver &resolver,
+                 std::string_view default_dialect = "");
   [[nodiscard]] auto begin() const -> const_iterator;
   [[nodiscard]] auto end() const -> const_iterator;
   [[nodiscard]] auto cbegin() const -> const_iterator;
@@ -67,7 +74,7 @@ private:
 #if defined(_MSC_VER)
 #pragma warning(disable : 4251)
 #endif
-  internal entries{};
+  internal subschemas{};
 #if defined(_MSC_VER)
 #pragma warning(default : 4251)
 #endif
