+++ test/foundation/foundation_test_utils.h
@@ -48,173 +48,4 @@
 #define TO_POINTER(pointer_string)                                             \
   sourcemeta::core::to_pointer((pointer_string))
 
-#define __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,        \
-                              expected_parent_location, expected_dialect,      \
-                              expected_base_dialect)                           \
-  EXPECT_EQ(sourcemeta::core::to_string(entries.at(index).pointer),            \
-            expected_schema_location);                                         \
-  EXPECT_OPTIONAL_POINTER(entries.at(index).parent, expected_parent_location); \
-  EXPECT_FALSE(entries.at(index).dialect.empty());                             \
-  EXPECT_TRUE(entries.at(index).base_dialect.has_value());                     \
-  EXPECT_EQ(entries.at(index).dialect, expected_dialect);                      \
-  EXPECT_EQ(                                                                   \
-      sourcemeta::blaze::to_string(entries.at(index).base_dialect.value()),    \
-      expected_base_dialect);                                                  \
-  EXPECT_FALSE(entries.at(index).vocabularies.empty());
-
-#define EXPECT_WALKER_ENTRY_2020_12(entries, index, expected_schema_location,  \
-                                    expected_parent_location)                  \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "https://json-schema.org/draft/2020-12/schema",        \
-                        "https://json-schema.org/draft/2020-12/schema");       \
-  EXPECT_FALSE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_2020_12_ORPHAN(                                    \
-    entries, index, expected_schema_location, expected_parent_location)        \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "https://json-schema.org/draft/2020-12/schema",        \
-                        "https://json-schema.org/draft/2020-12/schema");       \
-  EXPECT_TRUE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_2019_09(entries, index, expected_schema_location,  \
-                                    expected_parent_location)                  \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "https://json-schema.org/draft/2019-09/schema",        \
-                        "https://json-schema.org/draft/2019-09/schema");       \
-  EXPECT_FALSE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_2019_09_ORPHAN(                                    \
-    entries, index, expected_schema_location, expected_parent_location)        \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "https://json-schema.org/draft/2019-09/schema",        \
-                        "https://json-schema.org/draft/2019-09/schema");       \
-  EXPECT_TRUE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_DRAFT7(entries, index, expected_schema_location,   \
-                                   expected_parent_location)                   \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-07/schema#",             \
-                        "http://json-schema.org/draft-07/schema#");            \
-  EXPECT_FALSE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_DRAFT7_ORPHAN(                                     \
-    entries, index, expected_schema_location, expected_parent_location)        \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-07/schema#",             \
-                        "http://json-schema.org/draft-07/schema#");            \
-  EXPECT_TRUE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_DRAFT6(entries, index, expected_schema_location,   \
-                                   expected_parent_location)                   \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-06/schema#",             \
-                        "http://json-schema.org/draft-06/schema#");            \
-  EXPECT_FALSE(entries.at(index).orphan);                                      \
-  EXPECT_FALSE(entries.at(index).property_name);
-
-#define EXPECT_WALKER_ENTRY_DRAFT6_ORPHAN(                                     \
-    entries, index, expected_schema_location, expected_parent_location)        \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-06/schema#",             \
-                        "http://json-schema.org/draft-06/schema#");            \
-  EXPECT_TRUE(entries.at(index).orphan);                                       \
-  EXPECT_FALSE(entries.at(index).property_name);
-
-#define EXPECT_WALKER_ENTRY_DRAFT6_PROPERTY_NAME(                              \
-    entries, index, expected_schema_location, expected_parent_location)        \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-06/schema#",             \
-                        "http://json-schema.org/draft-06/schema#");            \
-  EXPECT_FALSE(entries.at(index).orphan);                                      \
-  EXPECT_TRUE(entries.at(index).property_name);
-
-#define EXPECT_WALKER_ENTRY_DRAFT4(entries, index, expected_schema_location,   \
-                                   expected_parent_location)                   \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-04/schema#",             \
-                        "http://json-schema.org/draft-04/schema#");            \
-  EXPECT_FALSE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_DRAFT4_ORPHAN(                                     \
-    entries, index, expected_schema_location, expected_parent_location)        \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-04/schema#",             \
-                        "http://json-schema.org/draft-04/schema#");            \
-  EXPECT_TRUE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_DRAFT3(entries, index, expected_schema_location,   \
-                                   expected_parent_location)                   \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-03/schema#",             \
-                        "http://json-schema.org/draft-03/schema#");            \
-  EXPECT_FALSE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_DRAFT3_ORPHAN(                                     \
-    entries, index, expected_schema_location, expected_parent_location)        \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-03/schema#",             \
-                        "http://json-schema.org/draft-03/schema#");            \
-  EXPECT_TRUE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_DRAFT2(entries, index, expected_schema_location,   \
-                                   expected_parent_location)                   \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-02/schema#",             \
-                        "http://json-schema.org/draft-02/hyper-schema#");      \
-  EXPECT_FALSE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_DRAFT2_ORPHAN(                                     \
-    entries, index, expected_schema_location, expected_parent_location)        \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-02/schema#",             \
-                        "http://json-schema.org/draft-02/hyper-schema#");      \
-  EXPECT_TRUE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_DRAFT1(entries, index, expected_schema_location,   \
-                                   expected_parent_location)                   \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-01/schema#",             \
-                        "http://json-schema.org/draft-01/hyper-schema#");      \
-  EXPECT_FALSE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_DRAFT1_ORPHAN(                                     \
-    entries, index, expected_schema_location, expected_parent_location)        \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-01/schema#",             \
-                        "http://json-schema.org/draft-01/hyper-schema#");      \
-  EXPECT_TRUE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_DRAFT0(entries, index, expected_schema_location,   \
-                                   expected_parent_location)                   \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-00/schema#",             \
-                        "http://json-schema.org/draft-00/hyper-schema#");      \
-  EXPECT_FALSE(entries.at(index).orphan);
-
-#define EXPECT_WALKER_ENTRY_DRAFT0_ORPHAN(                                     \
-    entries, index, expected_schema_location, expected_parent_location)        \
-  __EXPECT_WALKER_ENTRY(entries, index, expected_schema_location,              \
-                        expected_parent_location,                              \
-                        "http://json-schema.org/draft-00/schema#",             \
-                        "http://json-schema.org/draft-00/hyper-schema#");      \
-  EXPECT_TRUE(entries.at(index).orphan);
-
 #endif
